feat: move decoding functions to network utils
